backend/requirements.txt
ultralytics>=8.4.41
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
prisma==0.11.0
python-dotenv==1.0.0
numpy==1.24.3
websockets==12.0
